feat: component implementation.
Information is fetch from 'db-data' and assigned into variables in the main component.

We use three 'course-card' tag component, and then the variables values are assigned to the 'course' property by using [course] which is also declared in the 'course-card' component. We use 'Input' decorator to indicate that 'course' variable value comes from the HTML, it's required and also we declare it along with its type.

Finally, we use the 'course' variable in the 'course-card' HTML file using the '{{}}' notation to get its value depending of the main variable for 'description' and 'longDescription' and also we indicate 'iconUrl' for the [src] property using property binding.