feat: Output decorator.
A new event was added and linked with angular '(click)' wich executes a fucntion, this one emmits an event which contains the current course.

The 'courseSelected' binding in the course-card listens to the event and then executes the 'onCourseSelected()' fucntion and recieves the event.

Finally, in the main app component the previous called function gets executed.