feat: @for, @empty and contextual variables.
Instead of the three individual course variables we get an array of all courses.

We use the for directive to loop through each value from the array, and using track to identify each element.

Also, we use the empty directive to display a message if there's no any elment in the iterable.

We also get use of the contextual variables such as index (ehich is passed to the TS file and catched with the Input decorator), count, even, odd, etc. we get those values using a let variable inside the for directive and then we use the data binding inside the'course-card' component 'child'.