refactor: ngIf directive.
We add an undefined object element to test the old if directive.

The ngIf syntax is used in the img tag, which calls a function from 'course-card' which returns true if both the object and its iconUrl exists, if exists it displays the image, otherwise, it will display the 'ng-template' which has the 'noImage' as reference variable.