feat: content projection along with ContentChild directive.
Some exercises to get along with the use of contentChild.

A new componente was created: 'course-image' and it was added to the ngModules from app.component.

In the father or main component the 'course-image' component was injected.

In the 'course-card' component the 'ng-content' directive was used for content projection for 'course-image'.

Finally, the actual content was put into the course-image html template.