feat: host binding.
We make the 'highlighted' binding variable for data binding and we send the 'true' value to the component.

After, an @Input decorator receives this 'highlighted' variable value associated, by default the 'isHighLighted' associated variable has a 'false' default value, but we've sent a 'true' one, so it's 'true'.

Finally, we make use of the @HostBinding to assign to the component's class the 'highlighted' class, but this will only happen if the cssClasses() method returns true, an here is where we get use of the 'isHighLighted' variable that was set using the data binding.