feat: @HostListener decorator.
We use the decorator to listen to mouse vents both over and leave, so that, this will change the 'isHighLighted' variable and turn on the 'highlighted' class, (it turns off the class as well).

Also, this emitts an event wich goes out the ts thanks to 'toggleHighLight' variable decorated with @Output, and so, the html template recieves that event with '(toggleHighLight)' and sends it to 'onToggle(event)' where we also pass the event.

Finally, the function component gets executed and recieves a boolean value.