feat: getting references using exportAs.
We've used exportAs to allow obtain a reference to the directive.

In the html template we did: #highlighter='hl', to get a reference to the directive and using the event binding with '(dblclick)' we call the toggle() method from the directive.

Finally, in the main app.component we get a directly reference as example to compare with exportAs, and we access to it by using '@ViewChild'; first we get access to CourseCardComponent but we indicate with 'read' that we actually want the 'HighLightedDirective' directive instead of the coponent reference.