feat: structural directives.
A new diective was created 'ngx-unless'.

The constuctor receives both the content to insert 'course-image' and also the position on the DOM where will be inserted.

When the directive is invoked using *ngxUnless, the ngxUnless setter will receive '!course.iconUrl' and it will evaluated if the content will be shown or not.

As we're implementing the opposite of ngIf, when bith are false, we use 'createEmbeddedView' to insert the 'templateRef'element into the 'viewContainer' position on the DOM.