style: styles and ::ng-deep.
Some styles wer moved from the main globally CSS file into the 'course-card' css file, just the corresponding.

Since we're getting use of the content projection the '.course-card .course-description' class is not reachable for the component, so, to achieve make changes to the styles file we get use of '::ng-deep', this breaks the encapsulation and says 'Apply this style even though inside of child components', so that, we're bypassing the encapsulation.