feat: :host-context().
We put into a new div the ng-template and a 'test div' to try 'host-context'.

This is useful, as it helps us to apply the 'background-color' property to components which has a parent with '.salmon-theme' class, if the previous is met then the style gets applied to '.course-card'.

Yeah, I know, it's wierd XD.