feat: encapsulation: ViewEncapsulation.
We've explored some of the options Angular has to offer when it comes to encapsulating the components, we have 3 approaches.

ShadowDom uses the native ShadowDOM from the browser to encapsulate the component ccontents and its css styles into a 'shadow root'.

Then, we have None which stands for no encapsulation, so plain css styles will be applied, and also globally, be aware of that.

Finally, the Emulated which is the default encapsulation and the most used, it allows us to use pseudo-special classes such as ':host-context()' or '::ng-deep'.