feat: DI, http GET, subscribe.
We use Dependency injection to get use of a 'HttpClient' class.

The parameters are set by using 'HttpParams' along with the set() method.

Finally, we make an HTTP GET request to the url + params we've defined before, and using a subscriber which is an observable, we'll wait until the results comes in, so then, we assign the 'courses' results to 'this.courses' vairable, this variable now is initialized using these request results instead of initialazing it using the values from 'db-data' file.