feat: implicit subscription.
We get use of the implicit subscription by declaring courses variable as 'courses$', so in that way we can simplify the request.

Inside the html, we get use of the 'async' pipe to automatically subscribe the observable, we put that in a ngIf directive in order to get an alias 'courses' and use it inside other elements.