feat: service implemented.
The logic for the http GET request was moved from 'app.component' to 'courses.services', the 'CoursesService' was included and hence, 'loadCourses' from the service was called. In this way, now the http GET reequest could be used inside other components as a service instead of repeating the same logic for each component.