feat: custom service and data modifications with PUT.
We've made an event binding using 'courseChanged' which executes the save() function along with an event.

The component got injected with 'CoursesService' and the save() functions gets executed, and using the injected dependency, calls the 'courses.service' custom service. Also it stays susbscribed to wait for the result from the service.

Then, the saveCourse() function gets executed, initializes the headers and finally, performs the PUT request. The resul will be pass to the subscriber.