feat: services - dependency injection manually.
Inside the 'courses.services' we got rid of the Injectable parameter which performs the DI, we've implemented it manually.

In the main 'app.component' component, we created a function which returns the instance of 'CoursesService' type an d also it receives the 'HttpClient' dependency. Then we add the 'providers' property to the @Component's object.

Inside it we declare an object, passing three properties; 'provide' corresponding to a unique 'Token' which identifies the dependency, we crreated a const 'COURSES_SERVICE' variable for this as a 'InjectionToken' of 'CoursesService' type; useFactory that corresponds to the function we've previosuly created; finally, we pass 'deps' which corresponds to the 'HttpClient' since the function will need this dependency.

Finally, we only add '@Inject(COURSES_SERVICE)' next to the 'private coursesService: CoursesService' DI in the constructor to indicate that it needs to use this dependency by using the token along with the @Inject decorator.