feat: custom pipe implementation.
'filter-by-category-pipe' was created as a custom pipe. It filters the course based on both the courses array and the category to filter by.

On 'course-card' we add ngFor by taking the courses array form the component, the category specified, and sending both of them to the 'filterByCategory' pipe.

Also, it was needed to add the pipe whitin the 'declarations' and 'exports' from the courses.module.

Finally, some lines whithin 'course-card' were commented to allows this exercise works correctly.