feat: i18n Alternative expressions.
Use of alternative expressions to show the different existing categories with the i18n support within the course-card component.

The pipeline was removed from the app.component.

The missing category from the first element array from dt-data was added.

The 'ng extract-i18n' command was executed and the pertinent file was updated.