feat: @defer - on interaction with parameters.
We make use of 'interaction' but with parameters, using a button for displaying the content. And using another parameter for 'prefetch on' to download the bundle when the user interacts with the input, and then when the user interacts with the button the bundle is alredy there to be dosplayed.