feat: @defer - custom triggers.
We make use of custom triggers, in this case both prefetch and display will be done when the user clicks on either prefetch or displat button.

Take in account that if we don¿t click on prefetch and instead we go directly for the display button, the prefetch will be done immediately in anyway.